BackEnd-Quantum/requirements.txt
alembic==1.7.6
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.12.3
blinker==1.4
click==8.1.8
cryptography==3.4.8
distlib==0.3.4
filelock==3.6.0
fastapi
git-filter-repo==2.34.0
greenlet==1.1.2
h11==0.14.0
html-sanitizer==2.4.4
httplib2==0.20.2
importlib-metadata==4.6.4
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
platformdirs==2.5.1
PyJWT==2.3.0
pyparsing==2.4.7
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlalchemy>=1.4
typing_extensions==4.12.2
uvicorn==0.34.0
playwright
zipp==1.0.0
python-multipart
debugpy
psycopg2-binary
bcrypt
pydantic_settings
jinja2
websockets
psutil
sentence-transformers
torch
stripe
matplotlib
pandas
reportlab
python-multipart 
numpy
qrcode
pillow
